+++ app/src/components/UnscheduledDrafts.js
@@ -1,7 +1,9 @@
 import React, { useContext } from 'react';
 import PostList from './PostList';
 import AdminLinks from './AdminLinks';
+import Loading from './common/Loading';
 import NewPostButton from './common/NewPostButton';
+
 import { useFetchUnscheduledPosts } from '../lib/hooks';
 
 import PostsContext from '../PostsContext';
@@ -11,16 +13,22 @@ export default function UnscheduledDrafts() {
 
 	const { unscheduled } = posts;
 
-	useFetchUnscheduledPosts(posts, postsDispatch);
+	var isLoading = useFetchUnscheduledPosts(posts, postsDispatch);
 
 	return (
 		<>
-			<PostList
-				className="unscheduledDrafts"
-				date={false}
-				posts={unscheduled}
-			/>
-			<NewPostButton unscheduled={true} />
+			{isLoading ? (
+				<Loading />
+			) : (
+				<>
+					<PostList
+						className="unscheduledDrafts"
+						date={false}
+						posts={unscheduled}
+					/>
+					<NewPostButton unscheduled={true} />
+				</>
+			)}
 			<AdminLinks />
 		</>
 	);